added objects mirage and thing in vred

+++ src/vred/bbox.hpp
@@ -13,28 +13,28 @@ class Bbox {
 
 public:
   Bbox() {}
-  Bbox(Vect& _p, double _depth, double _width, double _height, double _alpha) : center(_p), depth(_depth), width(_width), height(_height), alpha(_alpha) {}
+  Bbox(Vect& _p, float _depth, float _width, float _height, float _alpha) : center(_p), depth(_depth), width(_width), height(_height), alpha(_alpha) {}
   virtual ~Bbox() {}
 
   Vect* rayIntersect(Ray& _ray) const;
   int  collide(const Bbox& box) const;
-  void set(const Vect& _center, double d, double h, double w, double a);
+  void set(const Vect& _center, float d, float h, float w, float a);
 
   Vect getCenter() const {return center;}
   void setCenter(const Vect& _center) {center = _center;}
-  double getHeight() const {return height;}
-  void setHeight(const double _height) {height = _height;}
-  double getWidth() const {return width;}
-  void setWidth(const double _width) {width = _width;}
-  double getDepth() const {return depth;}
-  void setDepth(double _depth) {depth = _depth;}
+  float getHeight() const {return height;}
+  void setHeight(const float _height) {height = _height;}
+  float getWidth() const {return width;}
+  void setWidth(const float _width) {width = _width;}
+  float getDepth() const {return depth;}
+  void setDepth(float _depth) {depth = _depth;}
   void calcGroup(int n, Solid** table);
   int  inner(Vect pt) const;
   void print();
 
 private:
   Vect center;
-  double depth, width, height, alpha;
+  float depth, width, height, alpha;
 
 };
